fix: dispose vjs when element unmounts
This commit fixes an issue where every time the element was
mounted/unmointed, the previous VJS instance would persist in the DOM.
This created clones of the VJS player and broke all the players inside
the custom element.

After this commit, the Lit CE will cleanup, `dispose` of, the VJS
instance whenever the component unmounts. Ensuring that in subsequent
renders the old VJS markup will not still be in the DOM.