feat: dynamically import vjs module
Before this commit:

Importing the vjs bundle was load-blocking for the rest of our code.
There was no way to tell the browser to load it and our custom element
async.

After this commit:

The vjs bundle is imported dynamically, so it doesn't block the browser
from loading the rest of the module.

While the majority of the code won't _execute_ until after vjs module
has loaded, it will have already been loaded. So the only bottle-neck
for execution becomes the player tech's bundle load time.

This opens the door for, in a future commit, allowing users to host
their own bundle of the player that's more aggressively minified than
what we can offer.

This also allows CDNs like unpckg to more aggressively bundle the split
deps.