feat: add build-url private method using js-core
This commit imports the js-coore package and uses it to build the video
source URL, including the `ixlib` params. It adds a `_buildURL` private
method to ix-video. This method accepts the source URL and any query
params, as an object. Then, it uses js-core to build said URL with the
params and returns the result. The result is then stored in the options
object as the source(s) source URL.