fix: only uopdate source in the updated hook
This commit addresses an issue where `this.source` was being updated
twice in the videojs lifecycle. First, when videojs was rendered for the
first time. It was given the complete setup object, including source.
Then, in the `updated()` hook, it was updating most of those option
values again. This happened in such quick succession that the initial
source URL request was getting canceled out by the updated source URL
set in the `updated()` hook.

In future, the logic behind setting the vjs option values needs to be
re-evaluated. As it stands, its very easy to inadvertently write to the
same videojs option / video property twice.